+++ oico/__main__.py
@@ -1,5 +1,4 @@
 from .cli.main import main
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # pragma: no cover
     main()
-